Break up plugins into individual config files
