+++ home/.config/nvim/lua/config/nvim-treesitter.lua
@@ -1,0 +1,9 @@
+return function()
+  require("nvim-treesitter.configs").setup({
+    ensure_installed = "maintained",
+    highlight = {
+      enable = true,
+      disable = {},
+    },
+  })
+end
