+++ home/.config/nvim/lua/config/vim-matchup.lua
@@ -1,0 +1,3 @@
+return function()
+  require("nvim-treesitter.configs").setup({ matchup = { enable = true } })
+end
